Split ```core.h``` into multiple files - Move non-template function definitions to .cc files - Split previos core.h into debug.h and instance.h
